Interactive plotting from any data source
Implemented a mechanism that allows the use in real-time of any type of
data source besides a data.frame (such as an SQL data base). This
required the redefinition of three argument function that together
define an interface for listing the database columns, the patient IDs
in the database, and the retrieval of the actual data for a (set of)
patient ID(s). This allows the transparent implementation of access to,
for example, a remote MySQL database, as described in the “database’
vignette…